+++ packages/client/src/WriteApi.ts
@@ -13,12 +13,17 @@ export default interface WriteApi {
    * Write lines of [Line Protocol](https://bit.ly/2QL99fu).
    *
    * @param lines - InfluxDB Line Protocol
+   * @param bucket
+   * @param org
+   * @param writeOptions
+   * @param timeout - timeout of the request
    */
   doWrite(
     lines: string[],
     bucket: string,
     org?: string,
-    writeOptions?: Partial<WriteOptions>
+    writeOptions?: Partial<WriteOptions>,
+    timeout?: number
   ): Promise<void>
 
   /**