+++ packages/client/src/impl/node/NodeHttpTransport.ts
@@ -125,12 +125,14 @@ export class NodeHttpTransport implements Transport {
    * @param body - message body
    * @param options
    * @param callbacks - communication callbacks
+   * @param timeout - timeout of the request
    */
   send(
     path: string,
     body: string,
     options: SendOptions,
-    callbacks?: Partial<CommunicationObserver<any>>
+    callbacks?: Partial<CommunicationObserver<any>>,
+    timeout?: number
   ): void {
     const cancellable = new CancellableImpl()
     if (callbacks && callbacks.useCancellable)
@@ -143,7 +145,8 @@ export class NodeHttpTransport implements Transport {
         this._request(message, cancellable, callbacks)
       },
       /* istanbul ignore next - hard to simulate failure, manually reviewed */
-      (err: Error) => callbacks?.error && callbacks.error(err)
+      (err: Error) => callbacks?.error && callbacks.error(err),
+      timeout
     )
   }
 
@@ -155,13 +158,15 @@ export class NodeHttpTransport implements Transport {
    * @param body
    * @param options - send options
    * @param responseStarted
+   * @param timeout - timeout of the request
    * @returns Promise of response body
    */
   request(
     path: string,
     body: any,
     options: SendOptions,
-    responseStarted?: ResponseStartedFn
+    responseStarted?: ResponseStartedFn,
+    timeout?: number
   ): Promise<any> {
     if (!body) {
       body = ''
@@ -172,53 +177,60 @@ export class NodeHttpTransport implements Transport {
     let contentType: string
     let responseStatusCode: number | undefined
     return new Promise((resolve, reject) => {
-      this.send(path, body as string, options, {
-        responseStarted(headers: Headers, statusCode?: number) {
-          if (responseStarted) {
-            responseStarted(headers, statusCode)
-          }
-          contentType = String(headers['content-type'])
-          responseStatusCode = statusCode
-        },
-        next: (data: Uint8Array): void => {
-          buffer = Buffer.concat([buffer, data])
-        },
-        complete: (): void => {
-          const responseType = options.headers?.accept ?? contentType
-          try {
-            if (responseStatusCode === 204) {
-              // ignore body of NO_CONTENT response
-              resolve(undefined)
+      this.send(
+        path,
+        body as string,
+        options,
+        {
+          responseStarted(headers: Headers, statusCode?: number) {
+            if (responseStarted) {
+              responseStarted(headers, statusCode)
             }
-            if (responseType.includes('json')) {
-              if (buffer.length) {
-                resolve(JSON.parse(buffer.toString('utf8')))
-              } else {
+            contentType = String(headers['content-type'])
+            responseStatusCode = statusCode
+          },
+          next: (data: Uint8Array): void => {
+            buffer = Buffer.concat([buffer, data])
+          },
+          complete: (): void => {
+            const responseType = options.headers?.accept ?? contentType
+            try {
+              if (responseStatusCode === 204) {
+                // ignore body of NO_CONTENT response
                 resolve(undefined)
               }
-            } else if (
-              responseType.includes('text') ||
-              responseType.startsWith('application/csv')
-            ) {
-              resolve(buffer.toString('utf8'))
-            } else {
-              resolve(buffer)
+              if (responseType.includes('json')) {
+                if (buffer.length) {
+                  resolve(JSON.parse(buffer.toString('utf8')))
+                } else {
+                  resolve(undefined)
+                }
+              } else if (
+                responseType.includes('text') ||
+                responseType.startsWith('application/csv')
+              ) {
+                resolve(buffer.toString('utf8'))
+              } else {
+                resolve(buffer)
+              }
+            } catch (e) {
+              reject(e)
             }
-          } catch (e) {
+          },
+          error: (e: Error): void => {
             reject(e)
-          }
+          },
         },
-        error: (e: Error): void => {
-          reject(e)
-        },
-      })
+        timeout
+      )
     })
   }
 
   async *iterate(
     path: string,
     body: string,
-    options: SendOptions
+    options: SendOptions,
+    timeout?: number
   ): AsyncIterableIterator<Uint8Array> {
     let terminationError: Error | undefined = undefined
     let nestedReject: (e: Error) => void
@@ -229,7 +241,14 @@ export class NodeHttpTransport implements Transport {
     const requestMessage = await new Promise<Record<string, any>>(
       (resolve, reject) => {
         nestedReject = reject
-        this._createRequestMessage(path, body, options, resolve, wrapReject)
+        this._createRequestMessage(
+          path,
+          body,
+          options,
+          resolve,
+          wrapReject,
+          timeout
+        )
       }
     )
     if (requestMessage.signal?.addEventListener) {
@@ -264,6 +283,10 @@ export class NodeHttpTransport implements Transport {
    *
    * @param path - API path starting with '/' and containing also query parameters
    * @param body - request body, will be utf-8 encoded
+   * @param sendOptions
+   * @param resolve
+   * @param reject
+   * @param timeout - timeout of the request
    * @returns a configuration object that is suitable for making the request
    */
   private _createRequestMessage(
@@ -271,7 +294,8 @@ export class NodeHttpTransport implements Transport {
     body: string,
     sendOptions: SendOptions,
     resolve: (req: http.RequestOptions) => void,
-    reject: (err: Error) => void
+    reject: (err: Error) => void,
+    timeout?: number
   ): void {
     const bodyBuffer = Buffer.from(body, 'utf-8')
     const headers: {[key: string]: any} = {
@@ -290,6 +314,7 @@ export class NodeHttpTransport implements Transport {
         ...headers,
         ...sendOptions.headers,
       },
+      timeout: timeout ?? this._defaultOptions.timeout,
     }
     if (sendOptions.signal) {
       options.signal = sendOptions.signal