+++ packages/client/src/transport.ts
@@ -24,12 +24,14 @@ export interface Transport {
    * @param requestBody - HTTP request body
    * @param options  - send options
    * @param callbacks - communication callbacks to received data in Uint8Array
+   * @param timeout - timeout of the request
    */
   send(
     path: string,
     requestBody: string,
     options: SendOptions,
-    callbacks?: Partial<CommunicationObserver<Uint8Array>>
+    callbacks?: Partial<CommunicationObserver<Uint8Array>>,
+    timeout?: number
   ): void
 
   /**
@@ -39,13 +41,16 @@ export interface Transport {
    * @param path - HTTP request path
    * @param requestBody - request body
    * @param options - send options
+   * @param responseStarted
+   * @param timeout - timeout of the request
    * @returns response data
    */
   request(
     path: string,
     requestBody: any,
     options: SendOptions,
-    responseStarted?: ResponseStartedFn
+    responseStarted?: ResponseStartedFn,
+    timeout?: number
   ): Promise<any>
 
   /**
@@ -55,11 +60,13 @@ export interface Transport {
    * @param path - HTTP request path
    * @param requestBody - request body
    * @param options - send options
+   * @param timeout - timeout of the request
    * @returns async iterable
    */
   iterate(
     path: string,
     requestBody: any,
-    options: SendOptions
+    options: SendOptions,
+    timeout?: number
   ): AsyncIterableIterator<Uint8Array>
 }